Board Version 2.5.3
1) I’ve included a full set of layers, but note that the bottom paste
layer is intentionally empty. This assumes that you will be reflowing
this board and probably using solder preforms on the through holes
parts. If you need a bottom paste layer, just let me know what you’d
like and I’ll whip one up!
2) I cleaned up the screens and so they do not touch any traces or pads.
3) I’ve added a note to the notes.txt indicating that the silk layers
are for reference and for test boards, and are not required on the
production version of the board.