Fixed bug in wdt_disable()!
Wholey crap that was a hard bug! The library routine wdt_disable()
incorrectly used an assignment to fiddle some watchdog bits, rather than
an OR. This had the effect of assigning 0 to the prescaler, which
sometimes could cause an immediate reset depending on the value of the
watchdog counter that that moment.  I replaced the library call with
inline C code that correctly fiddles the bits with an OR.