New Board Version 2.7.0
1. **Moved mounting holes **to accommodate new shell configuration.

2. **Eliminated parts R1 and R2**. These were there to protect the LEDs
from being overdriven in software. Because we are not allowing user
uploaded code, these are no longer necessary. LEDs are now driven
directly with very low duty PWM which saves parts and power..

3. **Eliminated parts R6 and D4**. These were there to allow the vibe to
be powered directly from the power jack without a battery present for
testing, but we turned out to not use this feature.

4. **Further increased the keep-out clearance around the ****through****
hole parts**. This takes advantage of some space we opened up by moving
the mounting holes and should give the factory more soldering process
options (bigger, less accurate selective soldering nozzle, or easier
hand soldering).

5. **Removed trace that allowed firmware to test the battery charging
chip**. We ended up functionally testing the battery charging chip, so
this feature is no longer needed and removing that trace increases
soldering clearance in a tight area.

6. **Simplified the programming-though-jack circuit**. The complexity
was not needed  since we will not be supporting Arduino-style code
uploads. We can still support adjusting parameters (like speeds and
patterns) though the jack- and we can decide if we want to include this
functionality in the shipping product anytime up until firmware code
freeze.